+++ src/index.js
@@ -12,12 +12,24 @@ import Onboarding from "./pages/Onboarding";
 import Create from "./pages/Create";
 import Challenge from "./pages/Challenge";
 import EditChallenge from "./pages/EditChallenge"
-import LearnView from "./pages/LearnView"
+import Learn from "./pages/Learn"
 
 import Leaderboards from "./pages/Leaderboards";
 import Ab from "./pages/Ab";
-import Learn from "./pages/Learn";
+
 import "./index.css";
+
+
+
+// LESSONS!!!
+import Ch1_1 from "./pages/content/Ch1_1"
+
+
+
+
+
+
+
 
 
 export default function App() {
@@ -40,7 +52,7 @@ export default function App() {
           <Route path="/practice/hard" element={<Practice type="hard" />} />
 
           <Route path="/learn" element={<Learn />} />
-          <Route path="/learn/:section/:activity" element={<LearnView />} />
+          <Route path="/learn/chapter1/lesson1" element={<Ch1_1 />} />
           <Route path="/challenges/:id" element={<Challenge />} />
           <Route path="/challenges/:id/edit" element={<EditChallenge />} />
           <Route path="/leaderboards/:id" element={<Leaderboards />} />
