--- src/pages/LearnView.js
@@ -1,331 +1,0 @@
-import { Link } from "react-router-dom";
-import { Fragment, useEffect, useRef, useState } from 'react'
-import { Disclosure, Menu, Transition, Dialog } from '@headlessui/react'
-import { BellIcon, MenuIcon, XIcon, FireIcon, ExclamationIcon, CheckIcon } from '@heroicons/react/outline'
-import { getAuth, onAuthStateChanged, signOut } from "firebase/auth";
-import { CheckCircleIcon } from '@heroicons/react/outline'
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-import { io } from "socket.io-client";
-import DashboardManager from "../modules/DashboardManager.js"
-import 'animate.css';
-import { Navigation } from '../components/navigation';
-
-const LearnView = () => {
-    const firebaseConfig = {
-        apiKey: "AIzaSyBLAN84VP3jSA5dqhrU6Bjmfu5NiUDuNw4",
-        authDomain: "cyberjags-8b081.firebaseapp.com",
-        databaseURL: "https://cyberjags-8b081.firebaseio.com",
-        projectId: "cyberjags-8b081",
-        storageBucket: "cyberjags-8b081.appspot.com",
-        messagingSenderId: "166652277588",
-        appId: "1:166652277588:web:e08b9e19916451e14dcec1",
-        measurementId: "G-7ZNKM9VFN2"
-    };
-
-    // Initialize Firebase
-    const app = initializeApp(firebaseConfig);
-    const analytics = getAnalytics(app);
-
-    const [Learn, setLearn] = useState({
-        data: []
-    })
-
-    const auth = getAuth();
-    //const socket = io("http://localhost:3002");
-
-
-    const [open, setOpen] = useState(true)
-    const [show, setShow] = useState(true)
-    const cancelButtonRef = useRef(null)
-    document.title = "CTFGuide - Dashboard"
-    const [user, setUser] = useState({
-        name: 'Loading...',
-        email: 'Loading...',
-        imageUrl: 'https://ctfguide.com/demopfp.png'
-    });
-
-    const [userData, setUserData] = useState({
-        streak: 0,
-        continueWorking: [],
-        username: "??",
-        points: 0
-    })
-
-
-    function logout() {
-        signOut(auth).then(() => {
-            // Sign-out successful.
-        }).catch((error) => {
-            // An error happened.
-        });
-
-    }
-    useEffect(() => {
-
-        var xhttp = new XMLHttpRequest();
-        xhttp.onreadystatechange = function() {
-            if (this.readyState === 4 & this.status === 200) {
-                console.log(this.responseText)
-                setLearn({
-                        data: JSON.parse(this.responseText)
-                    }
-                );
-
-            }
-        }
-        xhttp.open("GET", `${process.env.REACT_APP_API_URL}/challenges/Learn/global`);
-        xhttp.send();
-
-
-        onAuthStateChanged(auth, (firebaseUser) => {
-            if (firebaseUser) {
-                const uid = firebaseUser.uid;
-                console.log(firebaseUser.photoURL);
-
-                if (firebaseUser.photoURL) {
-                    setUser({
-                      name: firebaseUser.displayName,
-                      email: firebaseUser.email,
-                      imageUrl: firebaseUser.photoURL,
-                    });
-                    document.getElementById("pfp1").src = firebaseUser.photoURL
-                  } else {
-                    setUser({
-                      name: firebaseUser.displayName,
-                      email: firebaseUser.email,
-                      imageUrl:
-                        `https://ui-avatars.com/api/?name=${firebaseUser.email}&background=random`,
-                    });
-                    document.getElementById("pfp1").src = `https://ui-avatars.com/api/?name=${firebaseUser.email}&background=random`
-                  }
-
-                var xhttp = new XMLHttpRequest();
-                xhttp.onreadystatechange = function() {
-                    if (this.readyState === 4 & this.status === 200) {
-                        let data = JSON.parse(this.responseText);
-                        document.getElementById("navPoints").innerHTML = data.points;
-
-                        setUserData({
-                            username: data.username,
-                            streak: data.streak,
-                            continueWorking: data.continueWorking,
-                            points: data.points
-                        })
-
-
-                        document.getElementById("loader").classList.add("hidden");
-
-
-                    }
-
-                    if (this.readyState === 4 & this.status === 500) {
-                        // User not registered via API
-                        var xhttp = new XMLHttpRequest();
-                        xhttp.onreadystatechange = function() {
-                            if (this.readyState === 4 & this.status === 200) {
-                                window.location.reload();
-                            }
-                        }
-                        xhttp.open("GET", `${process.env.REACT_APP_API_URL}/users/register?uid=${firebaseUser.uid}`);
-                        xhttp.send();
-
-                    }
-                }
-
-                xhttp.open("GET", `${process.env.REACT_APP_API_URL}/users/data?uid=${firebaseUser.uid}`);
-                xhttp.send();
-                /*
-
-                      socket.emit('online', {
-                        uid: firebaseUser.uid,
-                      });
-
-                      setInterval(() => {
-                        socket.emit('heartbeat', {
-                        uid: firebaseUser.uid,
-                      })}
-                      , 5000);
-                      */
-
-            } else {
-                window.location.href = "../login";
-            }
-        });
-
-
-    }, []);
-
-
-    const steps = [
-        { id: '', name: 'Introduction', href: '#', status: 'learning' },
-        { id: '', name: 'Review Activity', href: '#', status: 'activity' },
-        { id: '', name: 'Using CTFGuide', href: '#', status: 'learning' },
-      ]
-    const people = [
-        {
-            name: 'Lindsay Walton',
-            imageUrl:
-                'https://images.unsplash.com/photo-1517841905240-472988babdf9?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=3&w=256&h=256&q=80',
-        },
-        // More people...
-    ]
-    const activityItems = [
-        { id: 1, person: people[0], project: 'Workcation', commit: '2d89f0c8', environment: 'production', time: '1h' },
-        // More items...
-    ]
-    const navigation = [
-        { name: 'Dashboard', href: '../dashboard', current: false },
-        { name: 'Practice', href: '../practice', current: false },
-        { name: 'Learn', href: './learn', current: false },
-        { name: 'Classes', href: '#', current: false },
-        { name: 'CTFLive', href: '#', current: false },
-        { name: 'Learn', href: '#', current: true },
-        { name: 'Friends', href: '#', current: false },
-    ]
-    const userNavigation = [
-        { name: 'Your Profile', href: '#' },
-        { name: 'Settings', href: '#' },
-        { name: 'Sign out', onClick: logout },
-    ]
-
-    function classNames(...classes) {
-        return classes.filter(Boolean).join(' ')
-    }
-
-    function dashboardTutorial() {
-        document.getElementById("dashboard_tutorial").classList.remove("hidden")
-    }
-    function dashboardTutorialDone() {
-        window.location.href = "../practice"
-    }
-
-    window.onload = function() {
-        if (localStorage.getItem("tutorial_phase") == 1) {
-            document.getElementById("dashboard_tutorial").classList.remove("hidden")
-        }
-
-        if (localStorage.getItem("tutorial_active")) {
-            document.getElementById("tutorial_banner").classList.add("hidden")
-        }
-    }
-
-    return (
-
-        <div className="min-h-full " style={{ fontFamily: 'Space Grotesk, sans-serif' }}>
-
-        <Navigation/>
-
-
-            <main className="mt-3" >
-
-            <div className="max-w-7xl mx-auto py-6 px-4 sm:px-6 lg:px-8   ">
-
-            <nav aria-label="Progress">
-      <ol role="list" className="border bg-gray-900 border-gray-700 rounded-md divide-y divide-gray-900 md:flex md:divide-y-0">
-        {steps.map((step, stepIdx) => (
-          <li key={step.name} className="relative md:flex-1 md:flex">
-            {step.status === 'complete' ? (
-              <a href={step.href} className="group flex items-center w-full">
-                <span className="px-6 py-4 flex items-center text-sm font-medium">
-                  <span className="flex-shrink-0 w-10 h-10 flex items-center justify-center bg-green-600 rounded-full ">
-                    <CheckIcon className="w-6 h-6 text-white" aria-hidden="true" />
-                  </span>
-                  <span className="ml-4 text-sm font-medium text-white">{step.name}</span>
-                </span>
-              </a>
-            ) : step.status === 'learning' ? (
-              <a href={step.href} className="px-6 py-4 flex items-center text-sm font-medium" aria-current="step">
-                <span className="flex-shrink-0 w-10 h-10 flex items-center justify-center border-2 border-white rounded-full">
-                  <span className="text-white"><i class="fas fa-book"></i> {step.id}</span>
-                </span>
-                <span className="ml-4 text-sm font-medium text-white">{step.name}</span>
-           </a>
-            ) : step.status === 'activity' ? (
-              <a href={step.href} className="px-6 py-4 flex items-center text-sm font-medium" aria-current="step">
-                <span className="flex-shrink-0 w-10 h-10 flex items-center justify-center border-2 border-white rounded-full">
-                  <span className="text-white"><i class="fas fa-pencil-alt"></i>  {step.id}</span>
-                </span>
-                <span className="ml-4 text-sm font-medium text-white">{step.name}</span>
-              </a>
-            ) : (
-              <a href={step.href} className="group flex items-center">
-                <span className="px-6 py-4 flex items-center font-medium">
-                  <span className="flex-shrink-0 w-10 h-10 flex items-center justify-center border-2 border-gray-300 rounded-full ">
-                    <span className="text-white ">{step.id}</span>
-                  </span>
-                  <span className="ml-4 text-2xl text-white ">{step.name}</span>
-                </span>
-              </a>
-            )}
-
-            {stepIdx !== steps.length - 1 ? (
-              <>
-                {/* Arrow separator for lg screens and up */}
-                <div className="hidden md:block absolute top-0 right-0 h-full w-5" aria-hidden="true">
-                  <svg
-                    className="h-full w-full text-gray-700"
-                    viewBox="0 0 22 80"
-                    fill="none"
-                    preserveAspectRatio="none"
-                  >
-                    <path
-                      d="M0 -2L20 40L0 82"
-                      vectorEffect="non-scaling-stroke"
-                      stroke="currentcolor"
-                      strokeLinejoin="round"
-                    />
-                  </svg>
-                </div>
-              </>
-            ) : null}
-          </li>
-        ))}
-      </ol>
-    </nav>
-
-              <div className="text-white">
-              <br></br>
-                <h1 className="text-4xl font-semibold ">Introduction</h1>
-                <p className="text-xl mt-3">
-                  Welcome to CTFGuide's Introduction to Cybersecurity course. This course is designed to teach you the basics of cybersecurity. At the end of this course, you'll have a solid understanding on what cybersecurity is, connect to servers, and solve basic CTF's.
-                </p>
-              </div>
-
-
-              <div className="text-white">
-              <br></br>
-                <h1 className="text-4xl font-semibold ">The CTFGuide Terminal</h1> 
-                <p className="text-xl mt-3">
-               For some challenges, you'll need to use a terminal. You can use the one built into your computer or use the one built in right here.
-               Our terminals have most security tools preinstalled and are ready to use.
-                  </p>
-
-                    
-                  <div id="terminal" className=" mt-6 ">
-                  <p className="text-gray-400 mb-2 hint"><span className="text-white ">Terminal (Beta)</span> Login as <span className="text-yellow-400">{userData.susername}</span> using the password <span className="text-yellow-400">{userData.spassword}</span><a style={{ cursor: 'pointer'}} className="hidden hover:bg-black text-gray-300">Need help?</a></p>
-                    <iframe className="w-full" height="500" src="https://terminal.ctfguide.com/wetty/ssh/root?pass=" ></iframe>
-                       </div>
-
-              </div>
-
-
-
-
-                </div>    
-                
-            </main>
-
-        
-        </div>
-
-
-    )
-
-
-}
-
-
-
-
-export default LearnView;
